# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: EasyConf Reviews Microservice
  description: "EasyConf is web-based conference management system, designed to organize\
    \ and administrate academic conferences, workshops, and events. It offers a platform\
    \ for authors, reviewers, and organizers (chairs) to handle paper submissions,\
    \ peer reviews, and program scheduling efficiently. It provides tools for managing\
    \ the review process, generating program schedules, and handling communication\
    \ between organizers and participants.\n\nThe Reviews Microservice handles the\
    \ process of reviewing papers before a conference. This microservice is responsible\
    \ for assigning papers to reviews, allowing reviewers to add their comments, and\
    \ allowing the PC chairs to finalize the decision about acceptance and rejection\
    \ of the submitted papers.\n\nThe different roles that the users can have are\
    \ specified in the Users Microservice and the process for submitting papers to\
    \ conferences is specified in the Submissions Microservice.\n\nTECHNICAL NOTE:\
    \ in every endpoint, we require the requesting user to provide their ID. I.e.,\
    \ the requesterID parameter must be provided. It is used for authentication in\
    \ order to check if the user is allowed to perform certain actions."
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /papers/{paperID}/bids:
    get:
      tags:
      - Bids
      summary: Get all bids for a given paper
      description: Responds with a list of bids and the IDs of the corresponding reviewers.
        The requester must be a chair of the track that the paper is in.
      operationId: getBidsForPaper
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Not found, the specified paper does not exist."
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successfully accessed the bids for paper
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidByReviewer'
        "403":
          description: "Forbidden, you are not allowed to view bids. The requester\
            \ must be a chair of the track."
    put:
      tags:
      - Bids
      summary: Bid on a paper
      description: "Saves the preference (based on expertise) of the requester in\
        \ regards to reviewing the given paper. The requester must be a reviewer of\
        \ the track the paper is in. Also, the submission deadline has to have passed\
        \ and the bidding deadline must not have passed."
      operationId: bid
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of the paper
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: Specifies a preference by the reviewer to review / not
                review a paper (based on their expertise)
              enum:
              - CAN_REVIEW
              - NEUTRAL
              - NOT_REVIEW
        required: true
      responses:
        "404":
          description: Not Found. The requested paper or track was not found.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "409":
          description: Conflict. The submission deadline must have passed. The bidding
            deadline must not have passed.
        "200":
          description: Successful bid on the paper
        "403":
          description: Forbidden. The requester must be a reviewer in the track the
            paper is in.
...
  /papers/{paperID}/title-and-abstract:
    get:
      tags:
      - Papers
      summary: Get the title and abstract of a paper
      description: 'Responds with the title and abstract of the given paper. '
      operationId: getTitleAndAbstract
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden. The requester must be a reviewer or chair for the
            track the paper is in.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "404":
          description: Not Found. The requested paper was not found.
        "200":
          description: Successful retrieval of paper title and abstract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperSummary'
  /papers/{paperID}/status:
    get:
      tags:
      - Papers
      summary: Get the review status of a paper
      description: "Responds with whether the paper has been accepted or rejected,\
        \ or if it hasn't been decided yet."
      operationId: getState
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to check status
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "404":
          description: Not Found. The requested paper was not found.
        "403":
          description: "Forbidden. The requester must be an author of the paper, a\
            \ chair of the paper's track, or a reviewer assigned to the paper."
        "200":
          description: Successful retrieval of the paper review status
          content:
            application/json:
              schema:
                type: string
                description: "The review status of the paper.\n\nACCEPTED: The paper\
                  \ has been accepted.\n\nREJECTED: The paper has been rejected.\n\
                  \nNOT_DECIDED: The review process has hot finished yet."
                enum:
                - ACCEPTED
                - REJECTED
                - NOT_DECIDED
...
  /papers/{paperID}/bids/by-reviewer/{reviewerID}:
    get:
      tags:
      - Bids
      summary: Get the bid of a given reviewer for a given paper
      description: "Responds with a bid (a preference to review or not review a paper,\
        \ based on expertise level, submitted by the reviewers) of the given reviewer\
        \ for reviewing the given paper. The requester must be the reviewer themselves,\
        \ or a chair of the track the paper is in."
      operationId: getBidForPaperByReviewer
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of the paper
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerID
        in: path
        description: The the ID of the reviewer of the paper
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: "Forbidden, you are not allowed to view bids. Only the chairs\
            \ of the tracks and the bid submitter are allowed to do that."
        "404":
          description: "Not found, the specified bid does not exist"
        "200":
          description: Successfully accessed the bids for paper by given reviewer
          content:
            application/json:
              schema:
                type: string
                description: Specifies a preference by the reviewer to review / not
                  review a paper (based on their expertise)
                enum:
                - CAN_REVIEW
                - NEUTRAL
                - NOT_REVIEW
...
components:
  schemas:
...
    PaperSummary:
      type: object
      properties:
        title:
          type: string
          description: The title of the paper
          example: Tails of frogs
        abstractSection:
          type: string
          description: The abstract of the paper
          example: Some sample abstract
      description: Includes the title and abstract of a paper for use during the Bidding
        process
...
    BidByReviewer:
      type: object
      properties:
        bidderID:
          type: integer
          description: The ID of the bidder
          format: int64
          example: 1
        bid:
          type: string
          description: Specifies a preference by the reviewer to review / not review
            a paper (based on their expertise)
          enum:
          - CAN_REVIEW
          - NEUTRAL
          - NOT_REVIEW
      description: The bid that a reviewer has made
...